app/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
huggingface-hub==0.23.0
humanize==4.9.0
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
ipyvue==1.11.1
ipyvuetify==1.9.4
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-text-splitters==0.2.1
langsmith==0.1.77
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
orjson==3.10.5
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymdown-extensions==10.8.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.1
pyzmq==26.0.3
reacton==1.8.3
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
responses==0.24.1
rich==13.7.1
rich-click==1.8.2
rpds-py==0.18.1
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
solara==1.33.0
solara-server==1.33.0
solara-ui==1.33.0
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
tenacity==8.4.1
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
yarl==1.9.4